requirements.txt
XlsxWriter>=1.2.2
azure-identity>=1.6.0
azure-mgmt-monitor>=2.0.0
azure-mgmt-redis>=12.0.0
azure-mgmt-resource>=18.0.0
azure-mgmt-subscription>=1.0.0
boto3>=1.17
openpyxl>=3.0.4
pandas>=1.3.0
pathlib>=1.0.1
redis>=3.5.3
